added inventory system with int for amount tracking
added an inventory system that followed the player and when the player collides with objects, they are moved into the "inventory" with an int for amount tracking shown in the bottom right corner of the inventory screen. Modified player code to be static and can be called in other scripts, making collision easier.